API documentation

API documentation

Index

FEMBase.Problem

JuliaFEM.Elasticity

Types

Missing docstring.

Missing docstring for JuliaFEM.Element. Check Documenter's build log for details.

FEMBase.Problem — Type.

problem(field_name, X, time)

Interpolate field from a set of elements defined in problem. Here, X is the location inside domain described by elements.

Internally, function loops through all the elements, finding the one containing the point X. After that, using inverse isoparametric mapping, first find dimensionless coordinates (ξ,η,ζ) of that element corresponding to the location of point X and after that interpolate the values of field under investigation. Algorithm can be expected to be somewhat slow for big models, but for tests models the performance is good.

Examples

Having a problem called body, one can query the field displacement at position X = (1.0, 2.0, 3.0) and time t = 1.0, with the command

X = (1.0, 2.0, 3.0)
time = 1.0
u = body("displacement", X, time)

source

JuliaFEM.Elasticity — Type.

Elasticity equations.

Field equation is:

m∂²u/∂t² = ∇⋅σ - b

Weak form is: find u∈U such that ∀v in V

δW := ∫ρ₀∂²u/∂t²⋅δu dV₀ + ∫S:δE dV₀ - ∫b₀⋅δu dV₀ - ∫t₀⋅δu dA₀ = 0

where

ρ₀ = density
b₀ = displacement load
t₀ = displacement traction

Formulations

plane stress, plane strain, 3D

References

https://en.wikipedia.org/wiki/Linearelasticity https://en.wikipedia.org/wiki/Finitestraintheory https://en.wikipedia.org/wiki/Stressmeasures https://en.wikipedia.org/wiki/Mooney%E2%80%93Rivlinsolid https://en.wikipedia.org/wiki/Strainenergydensityfunction https://en.wikipedia.org/wiki/Planestress https://en.wikipedia.org/wiki/Hooke'slaw

source

Functions

Missing docstring.

Missing docstring for JuliaFEM.update!. Check Documenter's build log for details.